requirements.txt
Flask
pandas
gspread
oauth2client
openai==0.27.1
python-dotenv
